requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.6
django-appconf==1.0.5
django-cacheops==7.0.2
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==23.3
djangorestframework==3.14.0
funcy==2.0
idna==3.6
pyodbc==5.0.1
python-dotenv==1.0.0
pytz==2022.7.1
redis==5.0.1
requests==2.31.0
sqlparse==0.4.3
typing_extensions==4.9.0
tzdata==2023.2
urllib3==2.2.0
